+++ src/wrap_cancel.rs
@@ -9,33 +9,40 @@ use crate::shutdown_signal::ShutdownSignal;
 /// The wrapped future is dropped when a shutdown is triggered before the future completes.
 /// The wrapped future is *not* dropped if it completes before the shutdown signal is received.
 #[must_use = "futures must be polled to make progress"]
-pub struct WrapCancel<F> {
-	pub(crate) shutdown_signal: ShutdownSignal,
-	pub(crate) future: Option<F>,
+pub struct WrapCancel<T: Clone, F> {
+	pub(crate) shutdown_signal: ShutdownSignal<T>,
+	pub(crate) future: Result<F, T>,
 }
 
-impl<F: Future> Future for WrapCancel<F> {
-	type Output = Option<F::Output>;
+impl<T: Clone, F: Future> Future for WrapCancel<T, F> {
+	type Output = Result<F::Output, T>;
 
 	#[inline]
 	fn poll(self: Pin<&mut Self>, context: &mut Context) -> Poll<Self::Output> {
 		// SAFETY: We never move `future`, so we can not violate the requirements of `F`.
 		let me = unsafe { self.get_unchecked_mut() };
 
+		// SAFETY: We never move `future`, so we can not violate the requirements of `F`.
+		// We do drop it, but that's fine.
 		match &mut me.future {
-			None => return Poll::Ready(None),
-			Some(future) => {
-				// SAFETY: We never move `future`, so we can not violate the requirements of `F`.
+			Err(e) => return Poll::Ready(Err(e.clone())),
+			Ok(future) => {
 				let future = unsafe { Pin::new_unchecked(future) };
 				if let Poll::Ready(value) = future.poll(context) {
-					return Poll::Ready(Some(value));
+					return Poll::Ready(Ok(value));
 				}
 			},
-		};
+		}
 
 		// Otherwise check if the shutdown signal has been given.
-		Pin::new(&mut me.shutdown_signal)
-			.poll(context)
-			.map(|()| None)
+		let shutdown = Pin::new(&mut me.shutdown_signal)
+			.poll(context);
+		match shutdown {
+			Poll::Ready(reason) => {
+				me.future = Err(reason.clone());
+				Poll::Ready(Err(reason))
+			},
+			Poll::Pending => Poll::Pending,
+		}
 	}
 }